+++ docs/changelog.rst
@@ -15,13 +15,17 @@ news
   var D and var D - var I, and lags 0 to 24
 - New graphs and table for each station of CrossCorrelation for
   var D and var D - var I, and lags -24 to 24
+- Frequency data (converted or not) label text inside of Correlation
+  graphs with which the data were processed
 - disable the analysis_the_best_periods_to_process because this
-  need more review and this need more theoretical support
+  need more review and this need more theoretical support (for
+  the future version)
 
 fixes
 +++++
 
 - Check the 'forecast_date' for bimonthly and trimonthly values
+- [code] Fixed return boolean in adjust_data_of_variables
 
 
 0.7.0 (**2013-11-13**)